Artist id must be an int

+++ config.ts
@@ -3,7 +3,7 @@ import { z } from "zod";
 const schema = z.object({
   BLUESKY_USERNAME: z.string(),
   BLUESKY_PASSWORD: z.string(),
-  GENIUS_ARTIST_ID: z.coerce.number(),
+  GENIUS_ARTIST_ID: z.coerce.number().int(),
   GENIUS_TOKEN: z.string(),
 });
